Adding 64 ste (MiniTool) + intent action to launch command at click
The idear is, an button permit to compute, defined by the parameters
from the MiniTool, the command to execute and we can see this on the
prreview tabs.

The fisrt solution was:
  At click for computing the node server ask to the python server by the
road "/compute". Unfortunately, i found that can't be done beacause of
if we define a road, the default road was erased.

  In consequence, i will execute shell command from the node server