feat(Processing): Replace installation commands
In order to streamline the installation process as well as improving
readability of the documentation. We advice users to install the default
dependencies then deciding to add, on top of it, CUDA dependencies.

It makes more sense to do this way instead of deliberately provoking an
error to see if user has a compatible GPU.

Added:

- `pnpm install:processing:cuda`
- `pnpm install:processing:cuda:windows`

Removed:

- `pnpm install:processing:cpu`
- `pnpm install:processing:cpu:windows`

BREAKING CHANGE: This commit introduces new installation process for the
`Processing` module. Now, CPU dependencies are installed first and by
default, then users can add CUDA dependencies on top by using the new
commands.