feat(Visualisation): Bundle visualisation services with Electron
This unusually big commit provides a new way for for accessing the
visualisation module.

By bundling the `front` and `audio` inside an Electron app, users will
be able to visualise data with minimal effort and installation onto
their computers.

The app is built and declined for ubuntu through a Debian package, a
`.dmg` for macOS > 10.15, and an self installing executable for windows
environments.

Each binary includes executables for `nodejs`, `ffmpeg`, and `ffprobe`,
allowing for the new audio service to be spawned within the Electron
app. (with express API to truncate audio before sending them through the
local network)

BREAKING CHANGE: The `front` and `audio` services are now bundled inside
an Electron app.